requirements.txt
flask
pyTelegramBotAPI
loguru
pillow
